+++ Serious-Engine/Sources/Engine/Sound/SoundObject.cpp
@@ -214,7 +214,7 @@ void CSoundObject::Play(CSoundData *pCsdLink, SLONG slFlags)
 
   //CPrintF("PLAY: '%s'", (const char*)pCsdLink->GetName().FileName());
   // get prediction tail
-  CSoundObject *psoTail = GetPredictionTail(EVENT_SOUNDPLAY, (ULONG)pCsdLink);
+  CSoundObject *psoTail = GetPredictionTail(EVENT_SOUNDPLAY, (ULONG)(size_t)pCsdLink);
   // if the event is predicted
   if (psoTail==NULL) {
     // do nothing;
@@ -350,7 +350,7 @@ void CSoundObject::Stop(void)
 
   CSoundObject *psoTail = this;
   // get prediction tail
-  psoTail = GetPredictionTail(EVENT_SOUNDSTOP, (ULONG)so_pCsdLink);
+  psoTail = GetPredictionTail(EVENT_SOUNDSTOP, (ULONG)(size_t)so_pCsdLink);
   // if the event is predicted
   if (psoTail==NULL) {
     // do nothing;
